Add loggingutils in test.support
